% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_tissue_score}
\alias{atac_tissue_score}
\title{ATAC-Seq Score for POL3 Genes Dataframe}
\format{
A dataframe with 9817 rows and 23 variables:
\describe{
\item{core_type}{Category for genes based on their expression across tissues}
\item{Chr}{Chromosome where the RNA of interest is located}
\item{Start}{Gene start coordinate}
\item{End}{Gene end coordinate}
\item{Gene}{Gene name}
\item{Index}{Unique RNA Sequence identifier - Can be retrieved from RNA Central}
\item{Type}{Type of RNA POL3 Transcript}
\item{Tissue}{Remaining columns contain the name of different assessed tissues}}
}
\source{
{Created by The VanBortle lab at UIUC to serve as an example}
}
\usage{
data('atac_tissue_score')
}
\description{
A list of tissues and the corresponding counts for RNA POL3 genes for each of them. Values represent a binary call on significance for accessibility. If the value is 1 the gene is accessible in the respective tissue.
}
\references{
Simon Lizarazo et al., 2025 bioRxiv Preprint
}
