% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnapol_counts}
\alias{rnapol_counts}
\title{RNA Polymerase raw counts dataframe}
\format{
A dataframe with 1061 rows and 7 variables:
\describe{
\item{Chr}{Chromosome where the gene is located}
\item{Start}{Gene start coordinate}
\item{Stop}{Gene stop coordinate}
\item{RNA_Type}{Type of transcript}
\item{pol1}{Raw counts for RNA Polymerase I Chip-Seq}
\item{pol2}{Raw counts  for RNA Polymerase II Chip-Seq}
\item{pol3}{Raw counts for RNA Polymerase III Chip-Seq}}
}
\source{
{Created by The VanBortle lab at UIUC to serve as an example}
}
\usage{
rnapol_counts
}
\description{
A subset of data obtained from CHIP-Seq Data for RNA Polymerase I, II and III
}
\keyword{datasets}
